--- src/main/java/me/dadus33/chatitem/packets/ChatItemPacket.java
@@ -1,58 +1,0 @@
-package me.dadus33.chatitem.packets;
-
-import org.bukkit.entity.Player;
-
-public class ChatItemPacket {
-
-	protected Player player;
-	protected Object packet;
-	protected PacketType type;
-	protected boolean cancel = false;
-	protected PacketMetadata metaData = new PacketMetadata();
-	
-	public ChatItemPacket(PacketType type, Object packet, Player player) {
-		this.player = player;
-		this.packet = packet;
-		this.type = type;
-	}
-	
-	public Player getPlayer() {
-		return player;
-	}
-
-	public boolean hasPlayer() {
-		return player != null;
-	}
-	
-	public String getPlayername() {
-		return getPlayer().getName();
-	}
-	
-	public Object getPacket() {
-		return packet;
-	}
-
-	public String getPacketName() {
-		return packet.getClass().getSimpleName();
-	}
-	
-	public PacketType getPacketType() {
-		return type;
-	}
-	
-	public boolean isCancelled() {
-		return cancel;
-	}
-	
-	public void setCancelled(boolean cancel) {
-		this.cancel = cancel;
-	}
-
-	public PacketContent getContent() {
-		return new PacketContent(this);
-	}
-	
-	public PacketMetadata getMetaData() {
-		return metaData;
-	}
-}
