--- src/main/java/me/dadus33/chatitem/packets/PacketHandler.java
@@ -1,7 +1,0 @@
-package me.dadus33.chatitem.packets;
-
-public abstract class PacketHandler {
-
-	public void onReceive(ChatItemPacket packet) {}
-	public void onSend(ChatItemPacket packet) {}
-}
