+++ src/main/java/me/dadus33/chatitem/utils/PacketUtils.java
@@ -12,7 +12,7 @@ public class PacketUtils {
 	public static final String NMS_PREFIX;
 	public static final String OBC_PREFIX;
 
-	public static final Class<?> CRAFT_PLAYER_CLASS, CRAFT_SERVER_CLASS, CHAT_SERIALIZER, COMPONENT_CLASS;
+	public static final Class<?> CRAFT_PLAYER_CLASS, CRAFT_SERVER_CLASS;
 	
 	/**
 	 * This Map is to reduce Reflection action which take more ressources than just RAM action
@@ -23,8 +23,6 @@ public class PacketUtils {
 		ALL_CLASS = new HashMap<>();
 		NMS_PREFIX = ProtocolVersion.getServerVersion().isNewerOrEquals(ProtocolVersion.V1_17) ? "net.minecraft." : "net.minecraft.server." + VERSION + ".";
 		OBC_PREFIX = "org.bukkit.craftbukkit." + VERSION + ".";
-		CHAT_SERIALIZER = getNmsClass("IChatBaseComponent$ChatSerializer", "network.chat.", "ChatSerializer");
-		COMPONENT_CLASS = getNmsClass("IChatBaseComponent", "network.chat.");
 		CRAFT_PLAYER_CLASS = getObcClass("entity.CraftPlayer");
 		CRAFT_SERVER_CLASS = getObcClass("CraftServer");
 	}
@@ -93,22 +91,6 @@ public class PacketUtils {
 			return null;
 		});
 	}
-	
-	/**
-	 * Get the current player ping
-	 * 
-	 * @param p the player
-	 * @return the player ping
-	 */
-	public static int getPing(Player p) {
-		try {
-			Object entityPlayer = getEntityPlayer(p);
-			return entityPlayer.getClass().getField("ping").getInt(entityPlayer);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		return 0;
-	}
 
 	/**
 	 * Get NMS player connection of specified player
@@ -139,51 +121,6 @@ public class PacketUtils {
 		try {
 			Object craftPlayer = CRAFT_PLAYER_CLASS.cast(p);
 			return craftPlayer.getClass().getMethod("getHandle").invoke(craftPlayer);
-		} catch (Exception e) {
-			e.printStackTrace();
-			return null;
-		}
-	}
-	
-	public static String getNmsEntityName(Object nmsEntity) {
-		try {
-			if(ProtocolVersion.getServerVersion().isNewerOrEquals(ProtocolVersion.V1_13)) {
-				Object chatBaseComponent = getNmsClass("Entity", "world.entity.").getDeclaredMethod("getDisplayName").invoke(nmsEntity);
-				return (String) getNmsClass("IChatBaseComponent", "network.chat.").getDeclaredMethod("getString").invoke(chatBaseComponent);
-			} else {
-				return (String) getNmsClass("Entity", "world.entity.").getDeclaredMethod("getName").invoke(nmsEntity);
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-			return null;
-		}
-	}
-
-	/**
-	 * Send the packet to the specified player
-	 * 
-	 * @param p which will receive the packet
-	 * @param packet the packet to sent
-	 */
-	public static void sendPacket(Player p, Object packet) {
-		try {
-			Object playerConnection = getPlayerConnection(p);
-			playerConnection.getClass().getMethod("sendPacket", getNmsClass("Packet", "network.protocol.")).invoke(playerConnection, packet);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-	
-	/**
-	 * Get NMS entity player of specified one
-	 * 
-	 * @param p the player that we want the NMS entity player
-	 * @return the entity player
-	 */
-	public static Object getCraftServer() {
-		try {
-			Object craftServer = CRAFT_SERVER_CLASS.cast(Bukkit.getServer());
-			return craftServer.getClass().getMethod("getHandle").invoke(craftServer);
 		} catch (Exception e) {
 			e.printStackTrace();
 			return null;
