+++ src/main/java/me/dadus33/chatitem/utils/Utils.java
@@ -3,25 +3,13 @@ package me.dadus33.chatitem.utils;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 
 public class Utils {
 
-	public static final String VERSION = Bukkit.getServer().getClass().getPackage().getName().replace(".", ",")
-			.split(",")[3];
-
 	public static String coloredMessage(String msg) {
 		return ChatColor.translateAlternateColorCodes('&', msg);
-	}
-
-	public static List<String> coloredMessage(String... messages) {
-		List<String> ret = new ArrayList<>();
-		for (String message : messages) {
-			ret.add(coloredMessage(message));
-		}
-		return ret;
 	}
 
 	public static List<Player> getOnlinePlayers() {
@@ -43,24 +31,5 @@ public class Utils {
 			e.printStackTrace();
 		}
 		return list;
-	}
-
-	/**
-	 * Get the current player ping
-	 * 
-	 * @param p the player
-	 * @return the player ping
-	 */
-	public static int getPing(Player p) {
-		try {
-			Object entityPlayer = PacketUtils.getEntityPlayer(p);
-			if(ProtocolVersion.getServerVersion().isNewerOrEquals(ProtocolVersion.V1_17))
-				return entityPlayer.getClass().getField("e").getInt(entityPlayer);
-			else
-				return entityPlayer.getClass().getField("ping").getInt(entityPlayer);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		return 0;
 	}
 }
